feat(app): rename model.ts→types.ts, add steiger FSD linter
- Rename entities/*/model.ts → types.ts: types live in types.ts,
  state/store stays model.ts if added later (clearer FSD intent)
- Install steiger + @feature-sliced/steiger-plugin; add lint:fsd script
- Fix fsd/no-public-api-sidestep: re-export queryKeys from shared/api
  index so entity hooks import from '../../shared/api' (not the
  segment file directly)
- steiger.config.ts: recommended rules with targeted overrides —
  insignificant-slice off (entities referenced from package root),
  no-segmentless-slices off for features (flat hooks files are
  deliberate), segments-by-purpose set to warn (hooks/types/schema
  are intentionally descriptive over canonical FSD names)

Run: bun --cwd packages/app lint:fsd
Result: 0 errors, 10 warnings (all segments-by-purpose on hook/type
segment names — acknowledged)
