+++ packages/app/src/entities/catalog/hooks.ts
@@ -1,6 +1,5 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { useApiClient } from '../../shared/api';
-import { queryKeys } from '../../shared/api/query-keys';
+import { queryKeys, useApiClient } from '../../shared/api';
 import { CatalogItemsResponseSchema } from './schema';
 
 type CatalogSortField =
