+++ packages/app/src/entities/catalog/index.ts
@@ -1,5 +1,5 @@
 export type { FetchCatalogItemsParams } from './api';
 export { fetchCatalogItems } from './api';
 export { useCatalogItemsInfinite } from './hooks';
-export type { CatalogItem, CatalogItemsResponse } from './model';
 export { CatalogItemSchema, CatalogItemsResponseSchema } from './schema';
+export type { CatalogItem, CatalogItemsResponse } from './types';
