+++ packages/app/src/entities/feed/index.ts
@@ -1,4 +1,4 @@
 export { fetchFeed } from './api';
 export { useFeed } from './hooks';
-export type { Comment, FeedResponse, Post } from './model';
 export { CommentSchema, FeedResponseSchema, PostAuthorSchema, PostSchema } from './schema';
+export type { Comment, FeedResponse, Post } from './types';
