+++ packages/app/src/entities/pack/index.ts
@@ -1,9 +1,9 @@
 export { fetchPack, fetchPacks } from './api';
 export { usePack, usePacks } from './hooks';
-export type { Pack, PackItem, PackListResponse, PackWithWeights } from './model';
 export {
   PackItemSchema,
   PackListResponseSchema,
   PackSchema,
   PackWithWeightsSchema,
 } from './schema';
+export type { Pack, PackItem, PackListResponse, PackWithWeights } from './types';
