+++ packages/app/src/entities/trip/index.ts
@@ -1,4 +1,4 @@
 export { fetchTrip, fetchTrips } from './api';
 export { useTrip, useTrips } from './hooks';
-export type { Trip, TripLocation } from './model';
 export { TripLocationSchema, TripSchema } from './schema';
+export type { Trip, TripLocation } from './types';
