+++ packages/app/src/entities/trip/types.ts
