+++ packages/app/src/entities/user/index.ts
@@ -1,4 +1,4 @@
 export { fetchCurrentUser, fetchUserProfile } from './api';
 export { useCurrentUser, useUserProfile } from './hooks';
-export type { User, UserProfile } from './model';
 export { UserProfileSchema, UserSchema } from './schema';
+export type { User, UserProfile } from './types';
