+++ packages/app/src/shared/api/index.tsx
@@ -22,3 +22,4 @@ export function useApiClient(): ApiClient {
 }
 
 export type { ApiClient };
+export * from './query-keys';
